enh: move the launchFullSync logic to a function that can be triggered with forceResync = true (sync) or false (unpause, resume)
